refactor to Gallary Component
